+++ packages/core/src/components/MapContainer.vue
@@ -33,11 +33,8 @@
           v-bind="moveHandle.actionButton.props"
         />
       </template>
-      <template v-if="moveHandle.closeContent" #closeContent>
-        <template v-if="typeof moveHandle.closeContent === 'string'">
-          {{ moveHandle.closeContent }}
-        </template>
-        <component :is="moveHandle.closeContent" v-else />
+      <template v-if="moveHandle.closeIcon" #closeIcon>
+        {{ moveHandle.closeIcon }}
       </template>
       <template #default>
         <component :is="moveHandle.component" v-bind="moveHandle.props || {}" />