+++ packages/plugins/Gfi/src/components/Gfi.vue
@@ -28,7 +28,6 @@ export default Vue.extend({
     ...mapGetters(['moveHandle']),
     ...mapGetters('plugin/gfi', [
       'actionButton',
-      'closeContent',
       'exportPropertyLayerKeys',
       'gfiContentComponent',
       'gfiConfiguration',
@@ -77,7 +76,9 @@ export default Vue.extend({
     },
     moveHandleProperties() {
       const properties: MoveHandleProperties = {
-        closeContent: this.closeContent,
+        closeIcon: this.gfiConfiguration.featureList
+          ? 'fa-angles-right'
+          : 'fa-xmark',
         closeLabel: t('plugins.gfi.header.close'),
         closeFunction: this.closeWindow,
         component: this.contentComponent,