+++ packages/plugins/Gfi/src/store/getInitialState.ts
@@ -1,9 +1,7 @@
 import { GfiState } from '../types'
-import CloseIcon from '../components/CloseIcon.vue'
 
 const getInitialState = (): GfiState => ({
   actionButton: null,
-  closeContent: CloseIcon,
   featureInformation: {},
   imageLoaded: false,
   visibleWindowFeatureIndex: 0,