+++ packages/plugins/Gfi/src/types.ts
@@ -13,7 +13,6 @@ import {
   RenderType,
   FeatureList,
   MoveHandleActionButton,
-  MoveHandleCloseContent,
 } from '@polar/lib-custom-types'
 
 /** parameter specification for request method */
@@ -31,7 +30,6 @@ export interface RequestGfiParameters {
 /** GFI Vuex Module State */
 export interface GfiState {
   actionButton: MoveHandleActionButton | null
-  closeContent: MoveHandleCloseContent
   /** default style for stroke and fill of the highlighted feature. */
   defaultHighlightStyle: HighlightStyle
   /** mapping of layer id to features found for last GFI call */